Move single-use methods to their respective steps
Move copy_if_changed from Step base class to ConfigureApplicationsStep
since it's only used there. Remove duplicate file_hash and files_match?
implementations from ConfigureFishStep to use the base class versions.
